+++ src/renderer/index.tsx
@@ -8,11 +8,11 @@ import { createRoot } from 'react-dom/client';
 import '@polkadot-cloud/core/accent/polkadot-relay.css';
 
 // App styles.
-import './theme/theme.css';
-import './theme/index.css';
+import './theme/theme.scss';
+import './theme/index.scss';
 
 // Library styles.
-import './library/Buttons/buttons.scss';
+import './library/kits/Buttons/buttons.scss';
 
 const rootElement = document.getElementById('root');
 if (!rootElement) {
