+++ src/renderer/library/kits/Buttons/ButtonSecondary/index.tsx
@@ -2,62 +2,51 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { motion } from 'framer-motion';
-import '@polkadot-cloud/core/css/buttons/ButtonSubmit/index.css';
-import type { ComponentBaseWithClassName } from 'types';
+import '@polkadot-cloud/core/css/buttons/ButtonSecondary/index.css';
 import type { ButtonCommonProps, ButtonIconProps } from '../types';
 import { onMouseHandlers } from '../Utils';
 import { appendOr, appendOrEmpty } from '@polkadot-cloud/utils';
+import type { ComponentBaseWithClassName } from '@/renderer/types';
 
-export type ButtonSubmitProps = ComponentBaseWithClassName &
+export type ButtonSecondaryProps = ComponentBaseWithClassName &
   ButtonIconProps &
   ButtonCommonProps & {
-    // use secondary network color.
-    colorSecondary?: boolean;
-    // button text.
-    text: string;
     // large button, small otherwise.
     lg?: boolean;
-    // pulsing effect.
-    pulse?: boolean;
+    // button text.
+    text: string;
   };
 
 /**
- * @name ButtonSubmit
- * @description Submit button style used within modals to submit transactions.
+ * @name ButtonSecondary
+ * @description Secondary button style used within the main interface of dashboards.
  */
-export const ButtonSubmit = ({
-  colorSecondary,
+export const ButtonSecondary = ({
   disabled,
   grow,
   iconLeft,
   iconRight,
   iconTransform,
+  lg,
   marginLeft,
   marginRight,
   marginX,
   className,
   style,
   text,
-  lg,
-  pulse,
   onClick,
   onMouseOver,
   onMouseMove,
   onMouseOut,
-}: ButtonSubmitProps) => (
-  <motion.button
-    whileHover={{ scale: !disabled ? 1.02 : 1 }}
-    whileTap={{ scale: !disabled ? 0.98 : 1 }}
-    className={`btn-submit${appendOr(lg, 'lg', 'sm')}${appendOrEmpty(
-      colorSecondary,
-      'secondary-color'
-    )}${appendOrEmpty(grow, 'grow')}${appendOrEmpty(marginRight, 'm-right')}${appendOrEmpty(
+}: ButtonSecondaryProps) => (
+  <button
+    className={`btn-secondary${appendOr(lg, 'lg', 'sm')}${appendOrEmpty(
+      grow,
+      'grow'
+    )}${appendOrEmpty(marginRight, 'm-right')}${appendOrEmpty(
       marginLeft,
       'm-left'
-    )}${appendOrEmpty(marginX, 'm-x')}${appendOrEmpty(pulse, 'pulse')}${
-      className ? ` ${className}` : ''
-    }`}
+    )}${appendOrEmpty(marginX, 'm-x')}${className ? ` ${className}` : ''}`}
     style={style}
     type="button"
     disabled={disabled}
@@ -78,5 +67,5 @@ export const ButtonSubmit = ({
         transform={iconTransform ? iconTransform : undefined}
       />
     ) : null}
-  </motion.button>
+  </button>
 );
