+++ src/renderer/library/kits/Buttons/ButtonText/index.tsx
@@ -4,7 +4,7 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { motion } from 'framer-motion';
 import '@polkadot-cloud/core/css/buttons/ButtonText/index.css';
-import type { ComponentBaseWithClassName } from 'types';
+import type { ComponentBaseWithClassName } from '@/renderer/types';
 import type { ButtonCommonProps, ButtonIconProps } from '../types';
 import { onMouseHandlers } from '../Utils';
 import { appendOrEmpty } from '@polkadot-cloud/utils';
