+++ src/renderer/screens/Home/Manage/Accounts.tsx
@@ -16,7 +16,7 @@ import type {
 } from '@/types/subscriptions';
 import { useManage } from './provider';
 import { getIcon } from '@/renderer/Utils';
-import { ButtonText } from '@/renderer/library/Buttons/ButtonText';
+import { ButtonText } from '@/renderer/library/kits/Buttons/ButtonText';
 
 export const Accounts = ({ setSection, setBreadcrumb, addresses }: AnyJson) => {
   const { getChainSubscriptions, getAccountSubscriptions, chainSubscriptions } =
