+++ src/renderer/screens/Import/Ledger/Manage.tsx
@@ -11,7 +11,7 @@ import { DragClose } from '../../../library/DragClose';
 import { AddressWrapper } from '../Addresses/Wrappers';
 import { Address } from './Address';
 import { determineStatusFromCodes } from './Utils';
-import { ButtonText } from '@/renderer/library/Buttons/ButtonText';
+import { ButtonText } from '@/renderer/library/kits/Buttons/ButtonText';
 
 export const Manage = ({
   addresses,
