+++ src/renderer/theme/theme.scss
@@ -18,7 +18,7 @@ SPDX-License-Identifier: GPL-3.0-only */
 
   --button-primary-background: #efeeee;
   --button-hover-background: #e8e6e6;
-  
+
   --modal-background-color: rgb(224 222 222 / 70%);
 
   --card-shadow-color: rgb(158 158 158 / 20%);
@@ -48,7 +48,7 @@ SPDX-License-Identifier: GPL-3.0-only */
 
   --button-primary-background: rgb(58 52 58);
   --button-hover-background: rgb(66 61 68);
-  
+
   --modal-background-color: rgb(22 18 22 / 70%);
 
   --card-shadow-color: rgb(28 24 28 / 25%);
@@ -59,4 +59,4 @@ SPDX-License-Identifier: GPL-3.0-only */
   --accent-color-stroke: var(--accent-color-stroke-dark);
   --accent-color-transparent: var(--accent-color-transparent-dark);
   --accent-color-pending: var(--accent-color-pending-dark);
-  }
+}
